requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
playwright==1.40.0
apscheduler==3.10.4
sendgrid==6.10.0
openai==1.3.5
python-multipart==0.0.6
aiohttp==3.9.1
beautifulsoup4==4.12.2
requests==2.31.0
pytest==7.4.3
httpx==0.25.2 